Refactor Web3Loader to use React.Context
This way we always just have one instance of Web3 being used.
