+++ stories/Web3Loader.tsx
@@ -1,16 +1,19 @@
 import { storiesOf } from "@storybook/react";
 import React from "react";
-import Web3Loader from "../src/components/Web3Loader";
+import { Web3Loader, Web3Provider } from "../src/components/Web3Context";
 
-storiesOf("Web3Loader", module)
-  .add("web3 is present", () => (
-    <Web3Loader ifPresent={() =>
-      <span> web3 is available </span>
-    }/>
-  )).add("web3 is not present", () => (
-    <Web3Loader ifNotPresent={() =>
-
-      <span> web3 is not available </span>
-    }/>
+storiesOf("Web3Context", module)
+  .add("render text ifPresent", () => (
+    <Web3Provider>
+      <Web3Loader ifPresent={() =>
+        <span>web3 is available</span>
+      }/>
+    </Web3Provider>
+  )).add("render text ifNotPresent", () => (
+    <Web3Provider>
+      <Web3Loader ifNotPresent={() =>
+        <span>web3 is not available</span>
+      }/>
+    </Web3Provider>
   ))
 ;
